accountflow-api/requirements.txt
Django==4.2.22
psycopg[binary]
djangorestframework==3.15.2
python-dotenv==1.0.0
PyJWT==2.8.0
django-jazzmin
pytest==8.4.2
Pillow>=11.0.0
pytest==8.4.2
djangorestframework
